Define an Operator type
