--- src/constants.tsx
@@ -1,9 +1,0 @@
-export const numbers = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9']
-export const operators = ['+', '-', '*', '/', '^', '%']
-
-export const status = {
-  green: 'green',
-  yellow: 'yellow',
-  gray: 'gray',
-  unguessed: 'unguessed',
-}
